+++ app/Infrastructure/Presistence/Eloquent/Repositories/EloquentUserRepository.php
@@ -3,6 +3,7 @@
 namespace App\Infrastructure\Presistence\Eloquent\Repositories;
 
 use App\Domain\User\Entities\User;
+use App\Domain\User\Exceptions\UserHasActiveRentalsException;
 use App\Domain\User\Repositories\UserRepository;
 use App\Infrastructure\Presistence\Eloquent\Models\UserModel;
 use Illuminate\Http\Request;
@@ -47,6 +48,14 @@ class EloquentUserRepository implements UserRepository
 
     public function delete(int $id): ?bool
     {
+        $hasActive = $this->model->newQuery()
+            ->where('id', $id)
+            ->hasActiveRental()
+            ->exists();
+
+        if ($hasActive) {
+            throw new UserHasActiveRentalsException();
+        }
         return $this->model->query()->where('id',$id)->delete();
     }
